chore: fix storybook setup for tables
